requirements.txt
PySide6-Essentials==6.8.0
pytest>=8.3.3
pywin32>=307; sys_platform == 'win32'
peewee
rarfile
